[HIB-188] Rename hy3p_worker -> worker in hyp3_daemon
